deprecating DataManager for DataDirectory.load(hint)
