+++ datatc/__init__.py
@@ -1,4 +1,5 @@
-from datatc.data_manager import DataManager
+from datatc.data_directory import DataDirectory
 from datatc.self_aware_data import SelfAwareData
+from datatc.data_manager import DataManager
 
-__all__ = ['DataManager', 'SelfAwareData']
+__all__ = ['DataDirectory', 'SelfAwareData', 'DataManager']
